+++ test/adapter/json/has_many_test.rb
@@ -2,7 +2,7 @@ require 'test_helper'
 
 module ActiveModel
   class Serializer
-    class Adapter
+    module Adapter
       class Json
         class HasManyTestTest < Minitest::Test
           def setup
