+++ test/adapter/json_api/belongs_to_test.rb
@@ -2,7 +2,7 @@ require 'test_helper'
 
 module ActiveModel
   class Serializer
-    class Adapter
+    module Adapter
       class JsonApi
         class BelongsToTest < Minitest::Test
           def setup
