+++ test/adapter/json_api/has_many_test.rb
@@ -2,7 +2,7 @@ require 'test_helper'
 
 module ActiveModel
   class Serializer
-    class Adapter
+    module Adapter
       class JsonApi
         class HasManyTest < Minitest::Test
           def setup
