+++ test/adapter/json_api/resource_type_config_test.rb
@@ -2,7 +2,7 @@ require 'test_helper'
 
 module ActiveModel
   class Serializer
-    class Adapter
+    module Adapter
       class JsonApi
         class ResourceTypeConfigTest < Minitest::Test
           def setup
